Rename queries/mutations and add authorization for some

+++ src/main/java/de/unistuttgart/iste/gits/course_service/controller/MembershipController.java
@@ -1,11 +1,14 @@
 package de.unistuttgart.iste.gits.course_service.controller;
 
+import de.unistuttgart.iste.gits.common.user_handling.LoggedInUser;
+import de.unistuttgart.iste.gits.common.user_handling.UserCourseAccessValidator;
 import de.unistuttgart.iste.gits.course_service.service.MembershipService;
 import de.unistuttgart.iste.gits.generated.dto.CourseMembership;
 import de.unistuttgart.iste.gits.generated.dto.CourseMembershipInput;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.graphql.data.method.annotation.Argument;
+import org.springframework.graphql.data.method.annotation.ContextValue;
 import org.springframework.graphql.data.method.annotation.MutationMapping;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.stereotype.Controller;
@@ -20,23 +23,38 @@ public class MembershipController {
 
     private final MembershipService membershipService;
 
-    @QueryMapping
+    @QueryMapping(name = "_internal_noauth_courseMembershipsByUserIds")
     public List<List<CourseMembership>> courseMembershipsByUserIds(@Argument final List<UUID> userIds) {
         return membershipService.getAllMembershipsByUserIds(userIds);
     }
 
     @MutationMapping
-    public CourseMembership createMembership(@Argument(name = "input") final CourseMembershipInput inputDto) {
+    public CourseMembership createMembership(@Argument(name = "input") final CourseMembershipInput inputDto,
+                                             @ContextValue final LoggedInUser currentUser) {
+        UserCourseAccessValidator.validateUserHasAccessToCourse(currentUser,
+                LoggedInUser.UserRoleInCourse.ADMINISTRATOR,
+                inputDto.getCourseId());
+
         return membershipService.createMembership(inputDto);
     }
 
     @MutationMapping
-    public CourseMembership updateMembership(@Argument(name = "input") final CourseMembershipInput inputDto) {
+    public CourseMembership updateMembership(@Argument(name = "input") final CourseMembershipInput inputDto,
+                                             @ContextValue final LoggedInUser currentUser) {
+        UserCourseAccessValidator.validateUserHasAccessToCourse(currentUser,
+                LoggedInUser.UserRoleInCourse.ADMINISTRATOR,
+                inputDto.getCourseId());
+
         return membershipService.updateMembershipRole(inputDto);
     }
 
     @MutationMapping
-    public CourseMembership deleteMembership(@Argument(name = "input") final CourseMembershipInput inputDto) {
+    public CourseMembership deleteMembership(@Argument(name = "input") final CourseMembershipInput inputDto,
+                                             @ContextValue final LoggedInUser currentUser) {
+        UserCourseAccessValidator.validateUserHasAccessToCourse(currentUser,
+                LoggedInUser.UserRoleInCourse.ADMINISTRATOR,
+                inputDto.getCourseId());
+
         return membershipService.deleteMembership(inputDto);
     }
 }
